% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_clocks.R
\name{all_clocks}
\alias{all_clocks}
\title{Run all methylation age calculations}
\usage{
all_clocks(
  x,
  coef,
  id_out = "ID",
  clocks = "age",
  cols_out = NULL,
  allow_missing = getOption("methylAge.allow_missing"),
  dim_warning = getOption("methylAge.dim_warning")
)
}
\arguments{
\item{x}{Matrix of methylation proportions, with probes/markers as named rows and samples as named columns.}

\item{coef}{data.frame of coefficients, with a column for marker and column for model coefficient.}

\item{id_out}{Name of the sample ID in the output tibble.}

\item{allow_missing}{If FALSE, then all markers in `coef` must be present in `x.`}

\item{dim_warning}{If TRUE, then a warning is raised when there are more columns than rows in `x`.}
}
\description{
Run all age clocks (or a selection of them).
Please cite the referenced articles for each clock if using this function.
}
\details{
This should be used on normalized data.

The clocks were developed in
}
\references{
\insertAllCited{}
}
