% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_methylation_data.R
\name{check_methylation_data}
\alias{check_methylation_data}
\title{Check methylation data is formatted in a way suitable for the methylAge package}
\usage{
check_methylation_data(x, dim_warning = TRUE)
}
\description{
Check methylation data is formatted in a way suitable for the methylAge package
}
