% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DunedinPoAm.R
\name{dunedinpoam_clean}
\alias{dunedinpoam_clean}
\title{Clean DunedinPoAm output to fit with the output of other estimators in the MethylAge package}
\usage{
dunedinpoam_clean(x, id_out = "ID", est_out = "DunedinPoAm")
}
\arguments{
\item{x}{Output from the DunedinPoAm38::PoAmProjector function.}

\item{id_out}{Column name for the sample ID column.}

\item{est_out}{Column name for the pace of ageing estimator.}
}
\description{
Reformat output from \insertCite{belsky2020quantification;textual}{methylAge} cortex tissueDNA methylation age calculation.
}
\details{
DunedinPoAm can be calculated from the DunedinPoAm38 package that can be installed from:
https://github.com/danbelsky/DunedinPoAm38
}
\references{
\insertAllCited{}
}
