% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DunedinPoAm.R
\name{dunedinpoam_helper}
\alias{dunedinpoam_helper}
\title{Convenience function for DunedinPoAm pace of ageing}
\usage{
dunedinpoam_helper(
  x,
  proportionOfProbesRequired = 0.8,
  id_out = "ID",
  est_out = "DunedinPoAm",
  allow_missing = getOption("methylAge.allow_missing"),
  dim_warning = getOption("methylAge.dim_warning")
)
}
\arguments{
\item{x}{Matrix or data.frame of methylation proportions, with probes/markers as `rownames` and samples as `colnames`.}

\item{proportionOfProbesRequired}{The threshold for missing data, as specified in DunedinPoAm38::PoAmProjector.}

\item{id_out}{Name of the sample ID in the output tibble.}

\item{est_out}{Name of the estimated pace of ageing length column in the output tibble.}

\item{allow_missing}{If FALSE, then all markers in `coef` must be present in `x.`}

\item{dim_warning}{If TRUE, then a warning is raised when there are more columns than rows in `x`.}
}
\description{
Calculate the \insertCite{belsky2020quantification;textual}{methylAge} cortex tissueDNA methylation age calculation.
Please cite the referenced article if using this function.
}
\details{
This relies on installing the package with
devtools::install_github("danbelsky/DunedinPoAm38")
}
\references{
\insertAllCited{}
}
