% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hannum.R
\name{hannum_clock}
\alias{hannum_clock}
\title{Hannum DNA methylation age calculation}
\usage{
hannum_clock(
  x,
  id_col = "ID",
  age_col = "hannum_mage",
  allow_missing = FALSE,
  dim_warning = TRUE
)
}
\description{
Hannum DNA methylation age calculation
}
