% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horvath.R
\name{horvath}
\alias{horvath}
\title{Horvath DNA methylation age calculation}
\usage{
horvath(
  x,
  id_col = "ID",
  age_col = "horvath_mAge",
  normalize = FALSE,
  dim_warning = TRUE
)
}
\description{
Horvath DNA methylation age calculation
}
