% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horvath.R
\name{horvath_clock_original}
\alias{horvath_clock_original}
\title{Horvath DNA methylation age calculation using original Horvath code}
\usage{
horvath_clock_original(
  x,
  id_out = "ID",
  age_out = "horvath_R_mage",
  normalize = FALSE,
  dim_warning = TRUE
)
}
\description{
Horvath DNA methylation age calculation using original Horvath code
}
