% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoage.R
\name{phenoage}
\alias{phenoage}
\title{Phenoage DNA methylation age calculation}
\usage{
phenoage(
  x,
  id_out = "ID",
  age_out = "phenoage",
  allow_missing = getOption("methylAge.allow_missing"),
  dim_warning = getOption("methylAge.dim_warning")
)
}
\arguments{
\item{x}{Matrix or data.frame of methylation proportions, with probes/markers as `rownames` and samples as `colnames`.}

\item{id_out}{Name of the sample ID in the output tibble.}

\item{age_out}{Name of the estimated age column in the output tibble.}

\item{allow_missing}{If FALSE, then all markers in `coef` must be present in `x.`}

\item{dim_warning}{If TRUE, then a warning is raised when there are more columns than rows in `x`.}
}
\description{
Calculate the \insertCite{phenoage;textual}{methylAge} DNA methylation Phenoage.
Please cite the referenced article if using this function.
}
\references{
\insertAllCited{}
}
