% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhang_en.R
\name{zhang_clocks}
\alias{zhang_clocks}
\title{Zhang DNA methylation age calculation}
\usage{
zhang_clocks(
  x,
  id_col = "ID",
  en_col = "zhang_en_mAge",
  blup_col = "zhang_blup_mAge",
  clocks = c("en", "blup"),
  dim_warning = TRUE
)
}
\description{
Zhang DNA methylation age calculation
}
