% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_clock.R
\name{generic_clock}
\alias{generic_clock}
\title{Generic linear model DNA methylation age calculation}
\usage{
generic_clock(
  x,
  coef,
  marker_col,
  coef_col,
  id_col = "ID",
  age_col = "mage",
  allow_missing = FALSE,
  dim_warning = TRUE,
  clock_name = "generic"
)
}
\description{
This function gives a methylation age based on the coefficients of a trained clock that
does not transform the data. Transforms may be performed outside of this function.
Several clocks in this package use this function for the clock calculation.
}
