% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_clock.R
\name{generic_clock}
\alias{generic_clock}
\title{Generic linear model DNA methylation age calculation}
\usage{
generic_clock(
  x,
  coef,
  col_marker = "marker",
  col_coef = "coefficient",
  intercept_name = "Intercept",
  id_out = "ID",
  age_out = "mage",
  allow_missing = getOption("methylAge.allow_missing"),
  dim_warning = getOption("methylAge.dim_warning"),
  clock_name = "generic"
)
}
\arguments{
\item{x}{Matrix or data.frame of methylation proportions, with probes/markers as `rownames` and samples as `colnames`.}

\item{coef}{data.frame of coefficients, with a column for marker and column for model coefficient.}

\item{col_marker}{Name of the marker/probe column in `coef`.}

\item{col_coef}{Name of the coefficient column in `coef`}

\item{intercept_name}{If character, the name of the intercept marker/probe in `coef`.
If numeric, then this is the intercept value.
If NULL, then the intercept is set to 0.}

\item{id_out}{Name of the sample ID in the output tibble.}

\item{age_out}{Name of the estimated age column in the output tibble.}

\item{allow_missing}{If FALSE, then all markers in `coef` must be present in `x.`}

\item{dim_warning}{If TRUE, then a warning is raised when there are more columns than rows in `x`.}

\item{clock_name}{Name of clock for errors and warnings.}
}
\value{
A tibble with and id column and estimated methylation age.
}
\description{
This function gives a methylation age based on the coefficients of a trained clock that
does not transform the data. Transforms may be performed outside of this function.
Several clocks in this package use this function for methylation age calculation.
}
\examples{

# Run the Hannum clock.
generic_clock(met, coef = as.data.frame(hannum_coef), intercept_name = NULL, clock_name = "Hannum")

# Run the Phenoage clock.
generic_clock(met, coef = phenoage_coef, clock_name = "Phenoage")

}
