% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horvath.R
\name{horvath_clock_original}
\alias{horvath_clock_original}
\title{Horvath DNA methylation age calculation using original Horvath code}
\usage{
horvath_clock_original(
  x,
  id_col = "ID",
  age_col = "horvath_mAge",
  normalize = FALSE,
  dim_warning = TRUE
)
}
\description{
Horvath DNA methylation age calculation using original Horvath code
}
