% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu_telomere.R
\name{lu_telomere}
\alias{lu_telomere}
\title{Lu DNA methylation telomere length (DNAmTL) calculation}
\usage{
lu_telomere(
  x,
  id_out = "ID",
  age_out = "DNAmTL",
  allow_missing = getOption("methylAge.allow_missing"),
  dim_warning = getOption("methylAge.dim_warning")
)
}
\arguments{
\item{x}{Matrix of methylation proportions, with probes/markers as named rows and samples as named columns.}

\item{id_out}{Name of the sample ID in the output tibble.}

\item{age_out}{Name of the estimated age column in the output tibble.}

\item{allow_missing}{If FALSE, then all markers in `coef` must be present in `x.`}

\item{dim_warning}{If TRUE, then a warning is raised when there are more columns than rows in `x`.}
}
\description{
Calculate the \insertCite{lu_telomere;textual}{methylAge} DNA methylation age.
Please cite the referenced article if using this function.
}
\references{
\insertAllCited{}
}
