% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu_telomere.R
\name{lu_telomere}
\alias{lu_telomere}
\title{Lu DNA methylation telomere length (DNAmTL) calculation}
\usage{
lu_telomere(
  x,
  id_col = "ID",
  age_col = "DNAmTL",
  allow_missing = FALSE,
  dim_warning = TRUE
)
}
\arguments{
\item{x}{Matrix of methylation proportions, with probes/markers as named rows and samples as named columns.}

\item{id_col}{Name of the sample ID in the output tibble.}

\item{age_col}{Name of the estimated age column in the output tibble.}

\item{allow_missing}{If FALSE, then all markers in `coef` must be present in `x.`}

\item{dim_warning}{If TRUE, then a warning is raised when there are more columns than rows in `x`.}
}
\description{
Lu DNA methylation telomere length (DNAmTL) calculation
}
